requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.0
charset-normalizer
colorama==0.4.6
django-cors-headers
decorator==4.4.2
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==1.0.3
mysqlclient
numpy==2.2.4
Pillow==11.1.0
proglog==0.1.10
PyJWT==2.10.1
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0